Verify endpoint certificates by default
Add a parameter to AdminClient to set certificate verification. This
should be always enabled! If using a private CA, either instantiate the
AdminClient with the path to the certificate or use the
REQUESTS_CA_BUNDLE environment variable.

It's annoying that requests does not use the system CA store, but this
is a whole other issue.